requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-anthropic
langgraph
guardrails-ai
python-dotenv
pydantic-settings
pyyaml
python-jose[cryptography]